Split up context tests
